requirements.txt
mediapipe
flask
numpy
joblib
pillow